requirements.txt
anthropic>=0.42.0
openai>=1.66.0
google-genai>=1.12.1
openpyxl>=3.1.5
sumolib>=1.24.0
pytest>=8.3.5
